SpaceDiceOld/pyproject.toml
[tool.poetry]
name = "spacedice"
version = "0.1.0"
description = ""
authors = ["E1000"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
markdown = "^3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
